WIP refactor node
