+++ dkg_engine/Cargo.toml
@@ -15,9 +15,8 @@ tokio = "1.20.1"
 hex = "0.4.2"
 
 #internal
-node = { path = "../node/core" }
-
-commands={path="../commands"}
-messages={path="../messages"}
-primitives={path="../primitives"}
+node = { path = "../node" }
+commands = { path = "../commands" }
+messages = { path = "../messages" }
+primitives = { path = "../primitives" }
 udp2p = { git = "https://github.com/ASmithOWL/udp2p", branch = "main" }
