+++ node/Cargo.toml
@@ -1,0 +1,69 @@
+[package]
+name = "node"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+primitives = { path = "../primitives" }
+block = { path = "../block" }
+telemetry = { path = "../telemetry" }
+vrrb-config = { path = "../vrrb-config" }
+commands = { path = "../commands" }
+network = { path = "../network" }
+messages = { path = "../messages" }
+lr_trie = { path = "../lr_trie" }
+state_trie = { path = "../state_trie" }
+tx_trie = { path = "../tx_trie" }
+storage = { path = "../storage" }
+miner = { path = "../miner" }
+claim = { path = "../claim" }
+ledger = { path = "../ledger" }
+wallet = { path = "../wallet" }
+txn = { path = "../txn" }
+state = { path = "../state" }
+validator = { path = "../validator" }
+events = { path = "../events" }
+reward = { path = "../reward" }
+blockchain = { path = "../blockchain" }
+patriecia = { git = "https://github.com/vrrb-io/patriecia", branch = "main" }
+trecho = { git = "https://github.com/vrrb-io/trecho", branch = "main" }
+udp2p = { git = "https://github.com/ASmithOWL/udp2p", branch = "main" }
+secp256k1 = { version = "0.20.2", features = ["rand"] }
+sha256 = "1.0.2"
+rand = { version = "0.6", features = ["std"] }
+bytebuffer = "0.2.1"
+chrono = "0.4.20"
+serde_json = "1.0.64"
+serde = { version = "1.0.101", features = ["derive"] }
+blake3 = "0.3.8"
+bip39 = "1.0.1"
+pickledb = "0.5.1"
+tokio = { version = "1.12.0", features = ["full"] }
+hex = "*"
+itertools = "0.10.1"
+thiserror = "1.0"
+ritelinked = { version = "0.3.2", features = ['serde'] }
+index_list = "0.2.7"
+bincode = "1.3.3"
+uuid = { version = "0.8", features = ["serde", "v4"] }
+tracing = "0.1.36"
+poem = "1.3.43"
+public-ip = "0.2.2"
+
+# log = "0.4.14"
+# strum = "0.21.0"
+# strum_macros = "0.21.0"
+# tracing = "0.1.36"
+# crossterm = { version = "0.19", features = ["serde"] }
+# tui = { version = "0.14", default-features = false, features = [
+#     'crossterm',
+#     'serde',
+# ] }
+# env_logger = "0.8.1"
+# ctrlc = "3.2.0"
+# simplelog = "0.10.0"
+# clipboard = "0.5.0"
+# futures = "0.3.1"
+# async-std = { version = "1", features = ["attributes", "tokio1"] }
