--- node/core/src/node/mod.rs
@@ -1,8 +1,0 @@
-mod node;
-mod node_auth;
-mod node_type;
-
-pub use node_auth::*;
-pub use node_type::*;
-
-pub use self::node::*;
