--- node/core/src/result.rs
@@ -1,12 +1,0 @@
-use thiserror::Error;
-
-#[derive(Debug, Error, Clone)]
-pub enum NodeError {
-    #[error("invalid node type {0} provided")]
-    InvalidNodeType(String),
-
-    #[error("{0}")]
-    Other(String),
-}
-
-pub type Result<T> = std::result::Result<T, NodeError>;
