--- node/runtime/Cargo.toml
@@ -1,75 +1,0 @@
-[package]
-name = "runtime"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "runtime"
-path = "src/lib.rs"
-
-[dependencies]
-clap = { version = "3.2.19", features = ["derive"] }
-telemetry = { path = "../../telemetry" }
-primitives = { path = "../../primitives" }
-storage = { path = "../../storage" }
-thiserror = "1.0.33"
-tokio = { version = "1.21.0", features = ["full"] }
-tracing = "0.1.36"
-tracing-subscriber = { version = "0.3.15", features = [
-    "fmt",
-    "registry",
-    "env-filter",
-    "tracing-log",
-    "json",
-] }
-secp256k1 = { version = "0.20.2", features = ["rand"] }
-sha256 = "1.0.2"
-rand = "0.8.4"
-uuid = { version = "0.8.0", features = ["serde", "v4"] }
-bytebuffer = "0.2.1"
-chrono = "0.4.19"
-serde_json = "1.0.64"
-serde = { version = "1.0.101", features = ["derive"] }
-blake3 = "0.3.8"
-bip39 = "1.0.1"
-pickledb = "0.4.1"
-futures = "0.3.1"
-async-std = { version = "1", features = ["attributes", "tokio1"] }
-env_logger = "0.8.1"
-hex = "*"
-itertools = "0.10.1"
-crossterm = { version = "0.19", features = ["serde"] }
-tui = { version = "0.14", default-features = false, features = [
-    'crossterm',
-    'serde',
-] }
-ctrlc = "3.2.0"
-simplelog = "0.10.0"
-log = "0.4.14"
-ritelinked = { version = "0.3.2", features = ['serde'] }
-strum = "0.21.0"
-strum_macros = "0.21.0"
-unicode-width = "0.1.9"
-clipboard = "0.5.0"
-public-ip = "0.2.1"
-udp2p = { git = "https://github.com/ASmithOWL/udp2p", branch = "main" }
-network = { path = "../../network" }
-node = { path = "../core" }
-commands = { path = "../../commands" }
-messages = { path = "../../messages" }
-events = { path = "../../events" }
-claim = { path = "../../claim" }
-pool = { path = "../../pool" }
-reward = { path = "../../reward" }
-ledger = { path = "../../ledger" }
-txn = { path = "../../txn" }
-wallet = { path = "../../wallet" }
-block = { path = "../../block" }
-state = { path = "../../state" }
-miner = { path = "../../miner" }
-blockchain = { path = "../../blockchain" }
-validator = { path = "../../validator" }
-anyhow = "1.0.64"
-tokio-test = "0.4.2"
-[features]
-node = []
