--- node/runtime/src/handler/mod.rs
@@ -1,3 +1,0 @@
-
-
-pub mod cmd;
