--- node/runtime/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod result;
-mod runtime;
-mod runtime_module;
-pub use runtime_module::*;
-
-pub use crate::runtime::*;
