--- node/runtime/src/runtime/mod.rs
@@ -1,7 +1,0 @@
-mod blockchain;
-mod miner;
-mod node_runtime;
-mod state;
-mod swarm;
-
-pub use node_runtime::*;
