+++ node/src/command_handler.rs
@@ -1,3 +1,20 @@
+use std::{net::SocketAddr, sync::mpsc::Sender};
+
+/// This module is the primary allocator in the system, it contains the data
+/// structures and the methods required to send commands to different parts of
+/// the system.
+use commands::command::Command;
+use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
+use udp2p::protocol::protocol::Message;
+
+struct MessageHeader {
+    //
+}
+
+struct MessageEnvelope {
+    //
+}
+
 /// The basic structure for allocating commands to different parts of the
 /// system.
 #[derive(Debug)]
@@ -9,13 +26,6 @@ pub struct CommandHandler {
     pub to_state_sender: UnboundedSender<Command>,
     pub to_gossip_tx: Sender<(SocketAddr, Message)>,
     pub receiver: UnboundedReceiver<Command>,
-}
-
-impl<T: Clone, V: Clone> MessageHandler<T, V> {
-    /// Creates and returns a new message handler.
-    pub fn new(sender: UnboundedSender<T>, receiver: UnboundedReceiver<V>) -> MessageHandler<T, V> {
-        MessageHandler { sender, receiver }
-    }
 }
 
 impl CommandHandler {
@@ -42,11 +52,12 @@ impl CommandHandler {
 
     /// Handles a command received by the command handler and allocates it to
     /// the proper part of the system for processing.
+    #[telemetry::instrument]
     pub fn handle_command(&mut self, command: Command) {
         match command {
             Command::StopMine => {
                 if let Err(e) = self.to_mining_sender.send(Command::StopMine) {
-                    println!("Error sending to mining sender: {:?}", e);
+                    telemetry::error!("Error sending to mining sender: {:?}", e);
                 }
             },
             Command::GetState => {
@@ -55,7 +66,7 @@ impl CommandHandler {
             },
             Command::ProcessTxn(txn) => {
                 if let Err(e) = self.to_mining_sender.send(Command::ProcessTxn(txn)) {
-                    println!(
+                    telemetry::error!(
                         "Error sending transaction to mining sender for processing: {:?}",
                         e
                     );
@@ -66,7 +77,7 @@ impl CommandHandler {
                     .to_mining_sender
                     .send(Command::ProcessTxnValidator(validator))
                 {
-                    println!(
+                    telemetry::error!(
                         "Error sending txn validator to mining sender for processing: {:?}",
                         e
                     );
@@ -74,7 +85,7 @@ impl CommandHandler {
             },
             Command::ProcessClaim(claim) => {
                 if let Err(e) = self.to_mining_sender.send(Command::ProcessClaim(claim)) {
-                    println!(
+                    telemetry::error!(
                         "Error sending new claim to mining receiver for processing: {:?}",
                         e
                     );
@@ -85,7 +96,7 @@ impl CommandHandler {
                     .to_mining_sender
                     .send(Command::StateUpdateCompleted(network_state))
                 {
-                    println!(
+                    telemetry::error!(
                         "Error sending updated network state to mining receiver: {:?}",
                         e
                     );
@@ -99,7 +110,7 @@ impl CommandHandler {
             },
             Command::SendMessage(src, message) => {
                 if let Err(e) = self.to_gossip_tx.send((src, message)) {
-                    println!("Error sending message command to swarm: {:?}", e);
+                    telemetry::error!("Error sending message command to swarm: {:?}", e);
                 }
             },
             Command::SendState(_requested_from, _lowest_block) => {},
@@ -109,7 +120,7 @@ impl CommandHandler {
                     component,
                     sender_id,
                 )) {
-                    println!(
+                    telemetry::error!(
                         "Error sending SendStateComponenets Command to state receiver: {:?}",
                         e
                     );
@@ -120,7 +131,7 @@ impl CommandHandler {
                     .to_state_sender
                     .send(Command::StoreStateComponents(data, component_type))
                 {
-                    println!(
+                    telemetry::error!(
                         "Error sending StoreStateComponentChunk to state receiver: {:?}",
                         e
                     );
@@ -132,13 +143,13 @@ impl CommandHandler {
                     .to_blockchain_sender
                     .send(Command::PendingBlock(block.clone(), sender_id))
                 {
-                    println!("Error sending pending block to miner: {:?}", e);
+                    telemetry::error!("Error sending pending block to miner: {:?}", e);
                 }
             },
             Command::InvalidBlock(_block) => {},
             Command::GetBalance(address) => {
                 if let Err(e) = self.to_mining_sender.send(Command::GetBalance(address)) {
-                    println!("Error sending GetBalance command to mining thread: {:?}", e);
+                    telemetry::error!("Error sending GetBalance command to mining thread: {:?}", e);
                 }
             },
             Command::SendGenesis(sender_id) => {
@@ -146,7 +157,7 @@ impl CommandHandler {
                     .to_blockchain_sender
                     .send(Command::SendGenesis(sender_id))
                 {
-                    println!(
+                    telemetry::error!(
                         "Error sending SendGenesis command to blockchain thread: {:?}",
                         e
                     );
@@ -155,16 +166,16 @@ impl CommandHandler {
             Command::MineGenesis => {},
             Command::GetHeight => {
                 if let Err(e) = self.to_blockchain_sender.send(Command::GetHeight) {
-                    println!(
+                    telemetry::error!(
                         "Error sending GetHeight command to blockchain thread: {:?}",
                         e
                     );
                 }
             },
             Command::MineBlock => {
-                info!("Received mine block command, starting the miner");
+                telemetry::info!("Received mine block command, starting the miner");
                 if let Err(e) = self.to_mining_sender.send(Command::StartMiner) {
-                    println!("Error sending Mine Block command to miner: {:?}", e);
+                    telemetry::error!("Error sending Mine Block command to miner: {:?}", e);
                 }
             },
             Command::ClaimAbandoned(sender_id, claim) => {
@@ -172,78 +183,65 @@ impl CommandHandler {
                     .to_mining_sender
                     .send(Command::ClaimAbandoned(sender_id, claim))
                 {
-                    println!("Error sending claim abandoned command to miner: {:?}", e)
+                    telemetry::error!("Error sending claim abandoned command to miner: {:?}", e)
                 }
             },
-            //TODO: Discuss whether we want to keep these, they have largely been outsourced to
-            // udp2p and will likely make more sense as part of the network specific
-            // modules.
-            //
-            // Command::Bootstrap(new_peer_addr, new_peer_pubkey) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::Bootstrap(new_peer_addr,
-            // new_peer_pubkey)) {         info!("Error sending bootstrap command to
-            // swarm: {:?}", e);     }
-            // }
-            // Command::AddNewPeer(new_peer_addr, new_peer_pubkey) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::AddNewPeer(new_peer_addr,
-            // new_peer_pubkey)) {         info!("Error sending add new peer command to
-            // swarm: {:?}", e);     }
-            // }
-            // Command::AddKnownPeers(data) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::AddKnownPeers(data)) {
-            //         info!("Error sending add known peers command to swarm: {:?}", e);
-            //     }
-            // }
-            // Command::AddExplicitPeer(peer_addr, peer_pubkey) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::AddExplicitPeer(peer_addr,
-            // peer_pubkey)) {         info!("Error sending add explicit peer command to
-            // swarm: {:?}", e);     }
-            // }
-            // Command::InitHandshake(peer_addr) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::InitHandshake(peer_addr)) {
-            //         info!("Error sending initialize handshake command to swarm: {:?}", e);
-            //     }
-            // }
-            // Command::ReciprocateHandshake(peer_addr, pubkey, signature) => {
-            //     if let Err(e) =
-            // self.to_swarm_sender.send(Command::ReciprocateHandshake(peer_addr, pubkey,
-            // signature)) {         info!("Error sending reciprocate handshake command
-            // to swarm: {:?}", e);     }
-            // }
-            // Command::CompleteHandshake(peer_addr, pubkey, signature) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::CompleteHandshake(peer_addr,
-            // pubkey, signature)) {         info!("Error sending complete handshake
-            // command to swarm: {:?}", e);     }
-            // }
-            // Command::ProcessAck(packet_id, packet_number, src) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::ProcessAck(packet_id,
-            // packet_number, src)) {         info!("Error sending process ack command
-            // to swarm: {:?}", e);     }
-            // }
-            // Command::CleanInbox(id) => {
-            //     if let Err(e) = self.to_swarm_sender.send(Command::CleanInbox(id.clone())) {
-            //         info!("Error sending process ack command to swarm: {:?}", e);
-            //     }
-            // }
             _ => {},
         }
     }
 }
 
-impl<T: Clone, V: Clone> Handler<T, V> for MessageHandler<T, V> {
-    fn send(&self, command: T) -> Option<T> {
-        if let Err(_) = self.sender.send(command.clone()) {
-            return None;
-        } else {
-            return Some(command);
-        }
-    }
-
-    fn recv(&mut self) -> Option<V> {
-        if let Ok(message) = self.receiver.try_recv() {
-            return Some(message);
-        } else {
-            return None;
-        }
-    }
-}
+//
+//TODO: Discuss whether we want to keep these, they have largely been
+// outsourced to udp2p and will likely make more sense as part of the network
+// specific modules.
+//
+// Command::Bootstrap(new_peer_addr, new_peer_pubkey) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::Bootstrap(new_peer_addr, new_peer_pubkey))
+// {         info!("Error sending bootstrap command to swarm: {:?}", e);     }
+// }
+// Command::AddNewPeer(new_peer_addr, new_peer_pubkey) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::AddNewPeer(new_peer_addr,
+// new_peer_pubkey)) {         info!("Error sending add new peer command to
+// swarm: {:?}", e);     }
+// }
+// Command::AddKnownPeers(data) => {
+//     if let Err(e) = self.to_swarm_sender.send(Command::AddKnownPeers(data)) {
+//         info!("Error sending add known peers command to swarm: {:?}", e);
+//     }
+// }
+// Command::AddExplicitPeer(peer_addr, peer_pubkey) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::AddExplicitPeer(peer_addr, peer_pubkey)) {
+// info!("Error sending add explicit peer command to swarm: {:?}", e);     }
+// }
+// Command::InitHandshake(peer_addr) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::InitHandshake(peer_addr)) {         info!
+// ("Error sending initialize handshake command to swarm: {:?}", e);     }
+// }
+// Command::ReciprocateHandshake(peer_addr, pubkey, signature) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::ReciprocateHandshake(peer_addr, pubkey,
+// signature)) {         info!("Error sending reciprocate handshake command
+// to swarm: {:?}", e);     }
+// }
+// Command::CompleteHandshake(peer_addr, pubkey, signature) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::CompleteHandshake(peer_addr,
+// pubkey, signature)) {         info!("Error sending complete handshake
+// command to swarm: {:?}", e);     }
+// }
+// Command::ProcessAck(packet_id, packet_number, src) => {
+//     if let Err(e) = self.to_swarm_sender.send(Command::ProcessAck(packet_id,
+// packet_number, src)) {         info!("Error sending process ack command
+// to swarm: {:?}", e);     }
+// }
+// Command::CleanInbox(id) => {
+//     if let Err(e) =
+// self.to_swarm_sender.send(Command::CleanInbox(id.clone())) {         info!("
+// Error sending process ack command to swarm: {:?}", e);     }
+// }
+//
