+++ node/src/lib.rs
@@ -1,0 +1,22 @@
+// TODO: remove once refactor is complete
+#![allow(warnings, unused)]
+
+pub mod command_handler;
+pub mod command_router;
+pub mod handler;
+pub mod message_handler;
+mod node;
+mod node_auth;
+mod node_type;
+pub mod result;
+mod runtime;
+mod runtime_module;
+pub(crate) mod test_utils;
+
+pub use node_auth::*;
+pub use node_type::*;
+pub use result::*;
+pub use runtime::*;
+pub use runtime_module::*;
+
+pub use crate::node::*;
