+++ node/src/node_auth.rs
