+++ node/src/node_type.rs
@@ -1,0 +1,1 @@
+pub use primitives::NodeType;
