+++ quorum/Cargo.toml
@@ -11,7 +11,7 @@ vrrb_vrf = { path = "../vrrb_vrf" }
 rand_chacha = "0.3.1"
 
 claim = { path = "../claim" }
-node = { path = "../node/core" }
+node = { path = "../node" }
 secp256k1 = { version = "0.20.2", features = ["rand"] }
 thiserror = "1.0.33"
 rand = { version = "0.6", features = ["std"] }
