+++ vrrb-config/Cargo.toml
@@ -1,17 +1,10 @@
 [package]
-name = "signer"
-version = "0.1.0"
+name = "vrrb-config"
+version = "0.0.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-hbbft = "0.1.1"
-thiserror = "1.0.32"
-
-serde = "1.0.144"
-
 primitives = { path = "../primitives" }
-node = { path = "../node" }
-dkg_engine = { path = "../dkg_engine" }
-
+secp256k1 = { version = "0.20.2", features = ["rand"] }
