+++ vrrb-config/src/lib.rs
@@ -1,0 +1,34 @@
+// use secp256k1::{PublicKey, Secp256k1, SecretKey};
+use std::{net::SocketAddr, path::PathBuf};
+
+#[derive(Debug, Clone)]
+pub struct NodeConfig {
+    pub id: primitives::NodeId,
+    pub idx: primitives::NodeIdx,
+    pub data_dir: PathBuf,
+    pub db_path: PathBuf,
+    pub node_idx: primitives::NodeIdx,
+    pub address: SocketAddr,
+    pub bootstrap: bool,
+    pub bootstrap_node_addr: SocketAddr,
+    pub node_type: primitives::NodeType,
+    // pub public_key: primitives::PublicKey,
+    // pub secret_key: primitives::SecretKey,
+}
+
+impl NodeConfig {
+    pub fn db_path(&self) -> &PathBuf {
+        &self.db_path
+    }
+
+    pub fn data_dir(&self) -> &PathBuf {
+        &self.data_dir
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+    #[test]
+    fn it_works() {}
+}
